shopping and auction system(similar to ebay)
This application is for providing online shopping functionality which is
divided in two component admin and users. This application will provide
bidding and auction system for selected products. User will be provided
a functionality to search and compare various products and their
sellers. Finally they can add products in their cart to proceed for
checkout. This will also keep record for all seller and customer of
certain product. ( Kept UI of this system very close to ebay )